Reviewer’s comments:
- remove additional external shuffle class
- added companion object and main method to the standalone shuffle class
- change to a shutdown hook instead of signal handlers
- various typos
