+++ core/src/main/scala/org/apache/spark/deploy/worker/ExternalShuffleService.scala
@@ -18,11 +18,14 @@
 package org.apache.spark.deploy.worker
 
 import org.apache.spark.{Logging, SparkConf, SecurityManager}
+import org.apache.spark.util.Utils
 import org.apache.spark.network.TransportContext
 import org.apache.spark.network.netty.SparkTransportConf
 import org.apache.spark.network.sasl.SaslRpcHandler
 import org.apache.spark.network.server.TransportServer
 import org.apache.spark.network.shuffle.ExternalShuffleBlockHandler
+
+import java.util.concurrent.CountDownLatch
 
 /**
  * Provides a server from which Executors can read shuffle files (rather than reading directly from
@@ -32,7 +35,7 @@ import org.apache.spark.network.shuffle.ExternalShuffleBlockHandler
  * Optionally requires SASL authentication in order to read. See [[SecurityManager]].
  */
 private[worker]
-class StandaloneWorkerShuffleService(sparkConf: SparkConf, securityManager: SecurityManager)
+class ExternalShuffleService(sparkConf: SparkConf, securityManager: SecurityManager)
   extends Logging {
 
   private val enabled = sparkConf.getBoolean("spark.shuffle.service.enabled", false)
@@ -62,5 +65,41 @@ class StandaloneWorkerShuffleService(sparkConf: SparkConf, securityManager: Secu
       server.close()
       server = null
     }
+  }
+}
+
+/**
+ * A main class for running the external shuffle service.
+ */
+object ExternalShuffleService extends Logging {
+  @volatile
+  private var server: ExternalShuffleService = _
+
+  private val barrier = new CountDownLatch(1)
+
+  def main(args: Array[String]): Unit = {
+    val sparkConf = new SparkConf
+    Utils.loadDefaultSparkProperties(sparkConf)
+    val securityManager = new SecurityManager(sparkConf)
+
+    // we override this value since this service is started from the command line
+    // and we assume the user really wants it to be running
+    sparkConf.set("spark.shuffle.service.enabled", "true")
+    server = new ExternalShuffleService(sparkConf, securityManager)
+
+    installShutdownHook()
+
+    // keep running until the process is terminated
+    barrier.await()
+  }
+
+  private def installShutdownHook() = {
+    Runtime.getRuntime.addShutdownHook(new Thread("External Shuffle Service shutdown thread") {
+      override def run() {
+        logInfo("Shutting down shuffle service.")
+        server.stop()
+        barrier.countDown()
+      }
+    })
   }
 }
