+++ docs/job-scheduling.md
@@ -101,7 +101,7 @@ then set `yarn.nodemanager.aux-services.spark_shuffle.class` to
 
 If you wish to use the external shuffle service in a Mesos deployment you need to launch the
 service on each Mesos slave that might run Spark jobs. A launcher script is provided in the `sbin/`
-directory: `bin/start-shuffle-service.sh`. You can later stop the service by running
+directory: `sbin/start-shuffle-service.sh`. You can later stop the service by running
 `sbin/stop-shuffle-service.sh`. Additional configuration options can be passed in `SPARK_SHUFFLE_OPTS`.
 
 ### Resource Allocation Policy
