+++ sbin/stop-shuffle-service.sh
@@ -22,4 +22,4 @@
 sbin="`dirname "$0"`"
 sbin="`cd "$sbin"; pwd`"
 
-"$sbin"/spark-daemon.sh stop org.apache.spark.shuffle.ExternalShuffleService 1
+"$sbin"/spark-daemon.sh stop org.apache.spark.deploy.worker.ExternalShuffleService 1
